fix edit rate

+++ database/migrations/2020_11_17_125928_create_rates_table.php
@@ -18,7 +18,7 @@ class CreateRatesTable extends Migration
             $table->string('name');
             $table->integer('time');
             $table->foreignId('department_id')->nullable()->constrained()->onDelete('cascade');
-            $table->double('rate')->nullable();
+            $table->decimal('rate')->nullable();
             $table->timestamps();
         });
     }
